[Perpetuals] refactor analysis and keep leverage
